correction branch condition

+++ Jenkinsfile
@@ -1,4 +1,4 @@
-if (env.BRANCH_NAME== 'release_1.0.0'){
+
 
 pipeline {
 
@@ -9,14 +9,17 @@ pipeline {
             stage("build") {
 
                 steps {
-                    sh 'docker system prune -af --volumes'
 
-                    sh 'echo building application !'
+                    script{
+                        if (env.BRANCH_NAME== 'release_1.0.0'){
+                            sh 'docker system prune -af --volumes'
 
-                    sh 'docker-compose up --build -d'
+                            sh 'echo building application !'
 
-                
-                    
+                            sh 'docker-compose up --build -d'
+
+                        }
+                    }
 
                 }
 
@@ -26,11 +29,14 @@ pipeline {
             stage("test") {
 
                 steps {
+                      script{
+                        if (env.BRANCH_NAME== 'release_1.0.0'){
 
-                   sh ' echo testing application'
-                   sh 'docker exec testjenkins2_develop_videgrenier_1 sh -c "./vendor/bin/phpunit ./tests"'
+                            sh ' echo testing application'
+                            sh 'docker exec testjenkins2_develop_videgrenier_1 sh -c "./vendor/bin/phpunit ./tests"'
                 
-                  
+                        }
+                    }
                 }
 
 
@@ -50,4 +56,3 @@ pipeline {
 
     }
     
-}
